reorganize imports for moved files
