+++ src/App.tsx
@@ -2,18 +2,18 @@ import React, { useEffect, useState, useCallback } from "react";
 import { TaskID, BaseState, Task, ROOT_ID } from "./domain/Task";
 import useTaskHooks from "./hooks/useTaskHooks";
 import { useTaskContext } from "./context/TaskContext";
-import TreeView from "./components/TreeView";
-import TodayUpcoming from "./components/TodayUpcoming";
-import CompletedToday from "./components/CompletedToday";
-import History from "./components/History";
-import DensityMenu from "./components/DensityMenu";
-import { AddTaskModal } from "./components/AddTaskModal";
-import { GracePeriodToast } from "./components/GracePeriodToast";
-import { SettingsPage } from "./components/SettingsPage";
-import { LiquidProgressAnimation } from "./components/LiquidProgressAnimation";
+import TreeView from "./components/tasks/TreeView";
+import TodayUpcoming from "./components/tasks/TodayUpcoming";
+import CompletedToday from "./components/tasks/CompletedToday";
+import History from "./components/tasks/History";
+import DensityMenu from "./components/shared/DensityMenu";
+import { AddTaskModal } from "./components/tasks/AddTaskModal";
+import { GracePeriodToast } from "./components/rewards/GracePeriodToast";
+import { SettingsPage } from "./components/settings/SettingsPage";
+import { LiquidProgressAnimation } from "./components/effects/LiquidProgressAnimation";
 import { useRewardsContext } from "./context/RewardsContext";
-import ThemeEffectsHost from "./components/ThemeEffectsHost";
-import AdvancedThreeEffectsHost from "./components/AdvancedThreeEffectsHost";
+import ThemeEffectsHost from "./components/effects/ThemeEffectsHost";
+import AdvancedThreeEffectsHost from "./components/effects/AdvancedThreeEffectsHost";
 import "./styles/app.css";
 
 const App: React.FC = () => {
