+++ src/components/tasks/TreeView.tsx
@@ -2,8 +2,8 @@
 import React, { useEffect, useState, useCallback, useMemo } from "react";
 import { css } from "@emotion/react";
 import TreeNode from "./TreeNode";
-import { Task, TaskID } from "../domain/Task";
-import useTaskHooks from "../hooks/useTaskHooks";
+import { Task, TaskID } from "../../domain/Task";
+import useTaskHooks from "../../hooks/useTaskHooks";
 
 interface TreeViewProps {
   rootTaskIds: TaskID[];
